refactor: rename `Navigation` to `nav`
